Update README.md
Just wanted to throw this out there - checking out the repo with the url provided puts me in the js-exploration-base branch by default, which doesn't contain the Vagrantfile, and breaks the original directions.  It took me a bit to realize my checked out copy was different than the displayed repo. The text above may or may not be the best replacement, but I offer it as a solution.